fix(cli,todo,deny): split monolith todo_cmd.rs, fix deny config, add split-project-docs skill
- Extract reference handlers from todo_cmd.rs (824→625 lines) to todo_ref_cmd.rs (205 lines)
- Fix cargo-deny: add xurl-core license clarification and git source allowlist
- Fix check-chinese: replace Chinese test string with ASCII equivalent
- Add split-project-docs skill for AGENTS.md-style documentation progressive disclosure
- Bump version to 0.1.97
